news_aggregator_recommendations/requirements.txt
Flask
pandas
sqlalchemy
scikit-learn
PyJWT
python-dotenv